requirements.txt
# Data
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2

# Time Series & ML
prophet==1.1.5
xgboost==2.0.3
tensorflow==2.16.1

# Reinforcement Learning
gymnasium==0.29.1
stable-baselines3==2.3.2

# API & App
fastapi==0.111.0
uvicorn==0.29.0
streamlit==1.35.0
anthropic==0.28.0

# Visualisation
matplotlib==3.9.0
seaborn==0.13.2
plotly==5.22.0

# Utilities
python-dotenv==1.0.1
pydantic==2.7.1
requests==2.32.2